requirements.txt
bidict>=0.21.2
certifi>=2020.12.5
chardet>=4.0.0
idna==2.10
python-engineio>=3.13.0
python-socketio>=4.6.0
requests>=2.25.1
six>=1.15.0
urllib3>=1.26.4
websocket-client>=0.57.0
openpyxl
pandas
python-dotenv
